chore: move directory from schema to api-clients
