+++ apps/web/src/api-clients/inventory/metric-data/composables/use-metric-data-api.ts
@@ -3,8 +3,8 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import type { AnalyzeResponse } from '@/api-clients/_common/schema/api-verbs/analyze';
 import type { ListResponse } from '@/api-clients/_common/schema/api-verbs/list';
 import type { MetricDataAnalyzeParameters } from '@/api-clients/inventory/metric-data/schema/api-verbs/analyze';
-import type { MetricDataListParameters } from '@/schema/inventory/metric-data/api-verbs/list';
-import type { MetricDataModel } from '@/schema/inventory/metric-data/model';
+import type { MetricDataListParameters } from '@/api-clients/inventory/metric-data/schema/api-verbs/list';
+import type { MetricDataModel } from '@/api-clients/inventory/metric-data/schema/model';
