+++ apps/web/src/api-clients/inventory/metric-data/schema/api-verbs/list.ts
