+++ apps/web/src/api-clients/inventory/metric-data/schema/model.ts
