+++ apps/web/src/api-clients/inventory/metric-example/schema/api-verbs/list.ts
