+++ apps/web/src/api-clients/inventory/metric-example/schema/model.ts
