+++ apps/web/src/lib/helper/config-data-helper.ts
@@ -5,7 +5,7 @@ import { cloneDeep, find } from 'lodash';
 import type { CostQuerySetModel } from '@/api-clients/cost-analysis/cost-query-set/schema/model';
 import type { DashboardModel } from '@/api-clients/dashboard/_types/dashboard-type';
 import type { WorkspaceModel } from '@/api-clients/identity/workspace/schema/model';
-import type { MetricExampleModel } from '@/schema/inventory/metric-example/model';
+import type { MetricExampleModel } from '@/api-clients/inventory/metric-example/schema/model';
 
 import type { DisplayMenu } from '@/store/display/type';
 import type { CloudServiceTypeReferenceMap } from '@/store/reference/cloud-service-type-reference-store';
