+++ apps/web/src/lib/variable-models/managed-model/resource-model/metric-data-variable-model.ts
@@ -1,4 +1,4 @@
-import type { MetricDataModel } from '@/schema/inventory/metric-data/model';
+import type { MetricDataModel } from '@/api-clients/inventory/metric-data/schema/model';
 
 import ResourceVariableModel from '@/lib/variable-models/_base/resource-variable-model';
 import type { VariableModelConstructorConfig } from '@/lib/variable-models/_base/types';
